+++ NEWS.md
@@ -1,4 +1,4 @@
-# textrecipes (development version)
+# textrecipes 1.1.0
 
 ## Improvements